update :: AuthUtil 로 코드 분리하기, 리펙토링
